change jwt operations to utils
