+++ app/api/v1/auth.py
@@ -1,8 +1,8 @@
 from app.schemas.common import LoginData
 import os
-from fastapi import APIRouter, Response, HTTPException, status
-from app.utils.authlogin import AuthMiddleware
+from fastapi import APIRouter, Response
 from app.providers.config import configs
+from app.utils.jwt import JWTUtils
 
 login = APIRouter()
 
@@ -10,8 +10,8 @@ login = APIRouter()
 @login.post("/")
 def login_user(response: Response, user: LoginData):
     if user.username == configs.username and user.password == configs.password:
-        auth =  AuthMiddleware(configs.secret_key, "HS256", "auth_token", configs.auth_server)
-        token = auth.create_access_token(data={"sub": user.username})
+        jwt_utils = JWTUtils(configs.secret_key, "HS256")
+        token = jwt_utils.create_access_token(data={"sub": user.username})
         response.set_cookie(
             key="auth_token",
             value=token,
